Factor out deployment test setup
