+++ configserver/src/test/apps/valivationOverride/services.xml
@@ -1,0 +1,4 @@
+<jdisc version="1.0">
+    <search/>
+    <nodes count="2"/>
+</jdisc>
